added getType
