+++ src/Devices/A4963/CMakeLists.txt
@@ -3,4 +3,4 @@ SET(LIBNAME a4963)
 file(GLOB SOURCE_FILES *.cpp)
 file(GLOB HEADER_FILES *.h)
 
-add_library(${LIBNAME} SHARED ${SOURCE_FILES} ${HEADER_FILES} A4963RegisterInfo.cpp)
+add_library(${LIBNAME} SHARED ${SOURCE_FILES} ${HEADER_FILES})
